backend/requirements.txt
# FastAPI 和相關套件
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic>=2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# 資料庫
pymongo>=4.13.0

# Docker (使用 6.x 版本以避免 urllib3 相容性問題)
urllib3==1.26.20
requests==2.32.5
requests-unixsocket==0.3.0
docker==6.1.3

# Git 操作
GitPython==3.1.41

# SSE (Server-Sent Events)
sse-starlette==1.8.2

# 工具
python-dotenv==1.0.0
python-multipart==0.0.6

# 認證
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# 測試
pytest==8.3.4
pytest-asyncio==1.3.0  # 需要 pytest>=8.2
httpx>=0.28.1
